requirements.txt
netmiko>=4.3.0
paramiko>=3.4.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
cryptography>=41.0.0
